diff/requirements.txt
aioredis
ccxt
